fix: close the pre-production E2E audit findings across modules
Findings come from the 2026-09-05 API-level E2E round (334 cases, 6 QA agents).

shared
- Central IntegrityError translation: unique -> 409 CONFLICTO_DUPLICADO, FK on
  insert -> 404 REFERENCIA_INEXISTENTE, FK on delete -> 409 EN_USO (was 500 in
  auth, turnos, prestadores and liquidaciones).
- Security headers middleware (nosniff, X-Frame-Options, Referrer-Policy,
  Permissions-Policy, Cache-Control no-store on /api, HSTS when cookie is Secure).
- /docs, /redoc and /openapi.json only in development; dev CORS origins only in
  development; SESSION_COOKIE_SECURE defaults to true outside development.
- uvicorn --proxy-headers --no-server-header; FORWARDED_ALLOW_IPS documented.

auth
- Login lockout keyed on the normalized email (letter-case variants bypassed it).
- Dummy argon2 verify for unknown emails and forgot-password mail sent via
  BackgroundTasks to stop user enumeration by timing.
- Forgot-password limited to 3 tokens per 15 min; inactive users get neither
  token nor reset; admin.manage grants reserved to superadmins; password
  max length 128; session ip/user agent persisted.
- Permissions/features: 404 for unknown user, 400 for unknown module/action/feature.

turnos
- Casilla: non-empty unique name. Slot: hora_inicio < hora_fin, dia 0..6, no
  overlap per casilla/day, 404 on unknown casilla/slot/user. Delete guards
  (slots in partial coverages, casillas with slots or active variants), 404 on
  unknown delete/put. Paginated /overrides, local date for variants, reads
  accept turnos.view or turnos.manage, GET /slots?fecha.

vacaciones / bono_tecnicos
- Suggested bonus days keep half-day granularity (no banker's rounding).
- Non-admins cannot request vacations starting before today (FECHA_PASADA).
- Superadmins receive the new-request mail; foreign empleadoIds -> 403.
- Attendance list flags weekend extension on day counts.

prestadores / contadores / preventivos
- 409 on duplicate siges id, 404 on unknown operator/user/PST; assignment
  history closes previous tramo and pointer only moves when desde <= today,
  reads resolve the operator by date; denComercial non-empty; single principal
  contact + email validation. Calendar dates typed, receso range validated and
  POST returns 201, clientes-nuevos bumps updated_at, unknown zone -> 404.

liquidaciones / analisis_log_hp
- Duplicate nombre_corto -> 409, SPST with unknown prestador -> 404, tarifario
  vigencia range validated. Saved analysis PUT persists name (non-empty) and
  health exposes the counters it used.

ci / frontend
- CI runs next build; Next emits the same security headers.

+++ backend/src/modules/bono_tecnicos/application/dtos/puntaje_tecnico_dto.py
@@ -21,7 +21,7 @@ class PuntajeTecnicoDTO:
     puntaje: float | None
     # None si el técnico no está vinculado a un empleado de Gestión de
     # Personal (`Empleado.siges_empresa_id`) — no hay de dónde sugerir.
-    dias_sugeridos: int | None
+    dias_sugeridos: float | None
 
 
 @dataclass(frozen=True, slots=True)
