+++ backend/src/modules/bono_tecnicos/domain/repositories/dias_sugeridos_gateway.py
@@ -6,7 +6,7 @@ from src.modules.bono_tecnicos.domain.value_objects.periodo import Periodo
 class DiasSugeridosGateway(Protocol):
     async def get_dias_sugeridos_por_tecnico(
         self, periodo: Periodo, ids_tecnico: list[int]
-    ) -> dict[int, int]:
+    ) -> dict[int, float]:
         """Días sugeridos por `id_tecnico` (días hábiles del período menos
         ausencias que reducen la jornada, ver `calculador_dias_sugeridos`)
         — solo para técnicos vinculados a un empleado de Gestión de
